test: move slow suites to ci tier
