+++ .github/workflows/ci.yml
@@ -39,6 +39,12 @@ jobs:
       - name: Run TypeScript checks
         run: pnpm run check
 
+      - name: Run slow TypeScript tests
+        run: pnpm run test:slow
+
+      - name: Run CLI smoke tests
+        run: pnpm run smoke
+
       - name: Setup Python
         uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
         with:
